+++ test/hexapdf/type/test_page_tree_node.rb
@@ -155,36 +155,44 @@ describe HexaPDF::Type::PageTreeNode do
     end
 
     it "deletes the correct page by index" do
-      assert_equal(@pages[2], @root.delete_page(2))
+      @root.delete_page(2)
       assert_equal(2, @kid12.page_count)
       assert_equal(4, @kid1.page_count)
       assert_equal(7, @root.page_count)
-      assert_nil(@pages[2][:Parent])
+      assert(@pages[2].null?)
 
-      assert_equal(@pages[5], @root.delete_page(4))
+      @root.delete_page(4)
       assert_equal(6, @root.page_count)
-      assert_nil(@pages[5][:Parent])
+      assert(@pages[5].null?)
     end
 
     it "deletes the given page" do
-      assert_equal(@pages[2], @root.delete_page(@pages[2]))
-      assert_equal(@pages[5], @root.delete_page(@pages[5]))
+      @root.delete_page(@pages[2])
+      assert(@pages[2].null?)
+      @root.delete_page(@pages[5])
+      assert(@pages[5].null?)
     end
 
     it "allows deleting a page from an intermediary node" do
-      assert_equal(@pages[2], @kid1.delete_page(@pages[2]))
+      @kid1.delete_page(@pages[2])
       assert_equal(7, @root.page_count)
     end
 
     it "does nothing if the page index is not valid" do
-      assert_nil(@root.delete_page(20))
-      assert_nil(@root.delete_page(-20))
+      @root.delete_page(20)
+      @root.delete_page(-20)
       assert_equal(8, @root.page_count)
     end
 
-    it "does nothing if the page is not in its parent's /Kids array" do
+    it "does nothing if the page has already been deleted" do
+      @root.delete_page(@pages[2])
+      @root.delete_page(@pages[2])
+      assert_equal(7, @root.page_count)
+    end
+
+    it "fails if the page is not in its parent's /Kids array" do
       @kid12[:Kids].delete_at(0)
-      assert_nil(@root.delete_page(@pages[2]))
+      assert_raises(HexaPDF::Error) { @root.delete_page(@pages[2]) }
       assert_equal(8, @root.page_count)
     end
 
@@ -193,7 +201,7 @@ describe HexaPDF::Type::PageTreeNode do
       page = @doc.add({Type: :Page, Parent: pages})
       pages[:Kids] << page
 
-      assert_nil(@root.delete_page(page))
+      @root.delete_page(page)
       assert_equal(8, @root.page_count)
     end
   end